+++ measure/src/main/scala/org/apache/griffin/measure/process/engine/SparkSqlEngine.scala
@@ -27,7 +27,6 @@ import org.apache.griffin.measure.process.temp.{DataFrameCaches, TableRegisters}
 import org.apache.griffin.measure.rule.adaptor.{GlobalKeys, InternalColumns}
 import org.apache.griffin.measure.rule.dsl._
 import org.apache.griffin.measure.rule.plan._
-import org.apache.griffin.measure.rule.step.TimeInfo
 import org.apache.griffin.measure.utils.JsonUtil
 import org.apache.spark.rdd.RDD
 import org.apache.spark.sql.{DataFrame, GroupedData, SQLContext}
