+++ measure/src/main/scala/org/apache/griffin/measure/rule/plan/TimeInfo.scala
@@ -16,12 +16,22 @@ KIND, either express or implied.  See the License for the
 specific language governing permissions and limitations
 under the License.
 */
-package org.apache.griffin.measure.rule.step
+package org.apache.griffin.measure.rule.plan
 
-import org.apache.griffin.measure.rule.dsl._
+trait TimeInfo extends Serializable {
+  val calcTime: Long
+//  val tmst: Long
+  val head: String
 
-case class DfOprStep(timeInfo: TimeInfo, ruleInfo: RuleInfo) extends ConcreteRuleStep {
-
-  val dslType: DslType = DfOprType
-
+  def key: String = if (head.nonEmpty) s"${head}_${calcTime}" else s"${calcTime}"
+  def setHead(h: String): TimeInfo
 }
+
+case class CalcTimeInfo(calcTime: Long, head: String = "") extends TimeInfo {
+//  val tmst: Long = calcTime
+  def setHead(h: String): TimeInfo = CalcTimeInfo(calcTime, h)
+}
+
+//case class TmstTimeInfo(calcTime: Long, tmst: Long, head: String = "") extends TimeInfo {
+//  def setHead(h: String): TimeInfo = TmstTimeInfo(calcTime, tmst, h)
+//}
